Using prettier to standardize the solidity code

+++ contracts/sale/ISaleFactory.sol
@@ -4,8 +4,7 @@ pragma solidity ^0.8.0;
 import "./ISaleData.sol";
 
 interface ISaleFactory {
-
-  event SaleApproved(uint saleId, address validator);
+  event SaleApproved(uint256 saleId, address validator);
   event NewSale(address saleAddress);
 
   function updateValidator(address validator) external;
@@ -14,12 +13,24 @@ interface ISaleFactory {
 
   function getSaleAddressById(uint256 i) external view returns (address);
 
-  function approveSale(uint saleId, ISaleData.Setup memory setup, ISaleData.VestingStep[] memory schedule, bytes memory signature) external;
+  function approveSale(
+    uint256 saleId,
+    ISaleData.Setup memory setup,
+    ISaleData.VestingStep[] memory schedule,
+    bytes memory signature
+  ) external;
 
-  function revokeApproval(uint saleId) external;
+  function revokeApproval(uint256 saleId) external;
 
-  function newSale(uint saleId, ISaleData.Setup memory setup, ISaleData.VestingStep[] memory schedule) external;
+  function newSale(
+    uint256 saleId,
+    ISaleData.Setup memory setup,
+    ISaleData.VestingStep[] memory schedule
+  ) external;
 
-   function encodeForSignature(uint saleId, ISaleData.Setup memory setup, ISaleData.VestingStep[] memory schedule) external pure returns (bytes32);
-
+  function encodeForSignature(
+    uint256 saleId,
+    ISaleData.Setup memory setup,
+    ISaleData.VestingStep[] memory schedule
+  ) external pure returns (bytes32);
 }
